directory hook no longer rechecks directory if it's same as the previous one

+++ src/Backend/fileHooks.js
@@ -69,6 +69,7 @@ export function useDirectory() {
 
   function setAbsDir(newPath) {
     const dirWorker = new Worker();
+    if (newPath === absDir) return;
     scanFiles(newPath, dirWorker).then(scanResult => {
       dirWorker.terminate();
       dispatch(handleNewFolder(newPath, scanResult))
